split streams
