+++ internal/server/events/main.go
@@ -1,4 +1,4 @@
-package server
+package events
 
 import (
 	"bytes"
