+++ makefile
@@ -1,11 +1,16 @@
+ui/node_modules:
+	@cd ui && npm install
+ui/dist:
+	@go generate ./...
+
 build:
 	@go generate ./...
 	@CGO_ENBALED=0 go build -ldflags="-w -s" -o ./bin/cli ./cmd/cli
 
-server:
+server: ui/dist
 	@go run cmd/cli/main.go server
 
-dev:
+dev: ui/node_modules
 	@cd ui && npm run dev
 
 docker-build:
@@ -14,4 +19,4 @@ docker-build:
 docker-run:
 	@docker run --rm --name svelte-server -p3000:3000 svelte-server:latest server -p 3000 --apiport 3001
 
-.PHONY: build server dev docker-build
+.PHONY: install-ui build server dev docker-build docker-run
