+++ ui/src/App.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
-  import BACKEND_BASE_URL from "$lib/config";
   import NotFound from "./routes/NotFound.svelte";
   import Router from "svelte-spa-router";
   import Squareterminal from "lucide-svelte/icons/square-terminal";
   import { wrap } from "svelte-spa-router/wrap";
   import { link } from "svelte-spa-router";
-  import { onDestroy, onMount } from "svelte";
-  import { subscribe, unsubscribe, es, connected } from "$lib/events";
+  import { connected } from "$lib/events";
+  import Wifi from "lucide-svelte/icons/wifi";
+  import Wifioff from "lucide-svelte/icons/wifi-off";
 
   const routes = {
     // Exact path
@@ -15,7 +15,7 @@
     }),
 
     "/memory": wrap({
-      asyncComponent: () => import("./routes/Homev2.svelte"),
+      asyncComponent: () => import("./routes/Memory.svelte"),
     }),
 
     "/syslog": wrap({
@@ -28,28 +28,6 @@
 
   let menuStyle =
     "text-muted-foreground hover:text-foreground transition-colors";
-
-  const connect = () => {
-    if (!$connected) {
-      es.set(new EventSource(`${BACKEND_BASE_URL}/stream`));
-      subscribe($es);
-    }
-  };
-
-  // Automatically subscribe when going to the page
-  onMount(() => {
-    connect();
-  });
-
-  // Automatically unsubscribe when leaving the page
-  onDestroy(() => {
-    unsubscribe($es);
-  });
-
-  // Handle user manually unsubscribing from server side events
-  function disconnect() {
-    unsubscribe($es);
-  }
 </script>
 
 <div class="flex min-h-screen w-full flex-col">
@@ -68,8 +46,12 @@
       <a href="/memory" use:link class={menuStyle}> Memory </a>
       <a href="/syslog" use:link class={menuStyle}> Syslog </a>
     </nav>
-    <div class="flex items-end gap-4">
-      <p>hello</p>
+    <div>
+      {#if $connected}
+        <Wifi />
+      {:else}
+        <Wifioff />
+      {/if}
     </div>
   </header>
   <Router {routes} />
