--- ui/src/lib/components/ui/separator/index.ts
@@ -1,7 +1,0 @@
-import Root from "./separator.svelte";
-
-export {
-	Root,
-	//
-	Root as Separator,
-};
