--- ui/src/routes/Homev2.svelte
@@ -1,9 +1,0 @@
-<script lang="ts">
-</script>
-
-<main class="bg-muted/40 flex flex-1 flex-col gap-4 p-4 md:gap-8 md:p-10">
-  <div class="mx-auto grid w-full max-w-6xl gap-2">
-    <h1 class="text-3xl font-semibold">Settings</h1>
-    <p>content</p>
-  </div>
-</main>
